Keep track of all trustees per device in a separate array
When someone creates a new trustee, always force a / trustee to be
created.  This trustee will service as the head of a doubly linked
list of all trustees created on that device.

This will be necessary to somewhat quickly iterate through all the
trustees in hopes of finding one that is a sub trustee of the current
path in the rename code.