Finally fixing major security bug with nested trustees
There are undoubtedly faster ways to do this, but this should be
correct at the very least.

When someone tries to rename a directory, we need to prevent
them from renaming a directory that has a trustee on a subpath.
Also, we need to prevent them from renaming *into* a directory
with a trustee on a subpath.

We simply return -XDEV which should force any program (mv) to
fall back to a copy + delete which is the only way to ensure
that the trustees are properly respected.