Changes to allow trustees to compile statically
The only notable non-makefile change is making trustees
initialize in the fs_initcall stage.  We really don't need to
start early in the initialization process and the major issue is
that securityfs initializes at core_initcall and the rest of the
security stuff initializes in security_initcall (before
securityfs!) I moved to fs_initcall to just avoid the issue all
together.